requirements.txt
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
Flask==2.2.2
google-api-core==2.11.0
google-api-python-client==2.72.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
markupsafe==2.1.1
oauthlib==3.2.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
twilio==7.16.1
uritemplate==4.1.1
urllib3==1.26.14
Werkzeug==2.2.2